Reorganize code tree for DUB registry.
- demo code is now under ./demo and has its own dub.json

  You can generate the demo VisualD project with:
  > cd demo
  > dub generate --compiler ldc2 [--combined] visuald

- source code of the library has moved to source/dg2d/ for import paths
- font.d now part of the library code instead of the example code.

What I would really recommend now is to have a package.d and only put there stuff that will be in the public API (in other words, boundaries).
It really helps long-term to separate stuff that is private from stuff that is in the public API (ideally: few things are public).

eg: dg2d.blitex would not be imported by package.d